+++ massgen/backend/utils/filesystem_manager/__init__.py
@@ -1,0 +1,29 @@
+"""Filesystem management utilities for MassGen backend."""
+from ._base import Permission
+
+from ._filesystem_manager import (
+    FilesystemManager
+)
+
+from ._path_permission_manager import (
+    ManagedPath,
+    PathPermissionManager,
+    PathPermissionManagerHook,
+)
+
+from ._workspace_copy_server import (
+    copy_file,
+    copy_files_batch,
+    get_copy_file_pairs,
+)
+
+__all__ = [
+    "FilesystemManager",
+    "ManagedPath",
+    "PathPermissionManager",
+    "PathPermissionManagerHook",
+    "Permission",
+    "copy_file",
+    "copy_files_batch",
+    "get_copy_file_pairs",
+]
