+++ massgen/backend/utils/filesystem_manager/_base.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+class Permission(Enum):
+    """File access permission types."""
+
+    READ = "read"
+    WRITE = "write"
