--- massgen/formatter/__init__.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Message formatting utilities.
-Provides utility classes for message formatting and conversion.
-"""
-from .mcp_tool_formatter import MCPToolFormatter
-from .message_formatter import MessageFormatter
-from .tool_formatter import ToolFormatter
-
-__all__ = ["MessageFormatter", "ToolFormatter", "MCPToolFormatter"]
